requirements.txt
python-docx==0.8.11
moviepy==1.0.3
requests==2.31.0
openai==1.3.0
Pillow==10.0.1
pydub==0.25.1
numpy==1.24.3 